+++ apps/client/src/environments/game-env.ts
@@ -1,5 +1,6 @@
 export const gameEnv = {
-  globalGameVersion: 7.5,
-  koreanGameVersion: 7.5,
-  chineseGameVersion: 7.5
+  globalGameVersion: 7.51,
+  koreanGameVersion: 7.51,
+  chineseGameVersion: 7.51,
+  traditionalChineseGameVersion: 7.11
 };