+++ libs/data/src/lib/model/lazy-fishing-spots-database-page.ts
@@ -16,7 +16,7 @@ export interface LazyFishingSpotsDatabasePage {
   radius:        number;
   tw?:           string;
   weatherRate:   number;
-  zh:            string;
+  zh?:           string;
   zoneId:        number;
 }